# services/manager/openapi/v1.yaml
openapi: 3.1.0
info:
  title: SkausWatch Manager API
  description: SkausWatch manager service — users, alerts, threat intelligence, approvals, ENDPOINT agent fleet, SIEM, ASM, CodeScan proxy, S3 malware scanning, and research lookups.
  contact:
    name: Penguin Tech Inc
    email: support@penguintech.io
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  version: '1'
paths:
  /api/v1/alerts:
    get:
      tags:
      - alerts
      summary: |-
        GET /alerts — paginated alert list, filterable by severity/status/source
        via repeated query params.
      operationId: list_alerts
      parameters:
      - name: page
        in: query
        description: 1-based page number (default 1)
        required: false
        schema:
          type: integer
          format: int64
      - name: per_page
        in: query
        description: Page size, capped at 100 (default 20)
        required: false
        schema:
          type: integer
          format: int64
      - name: severity
        in: query
        description: Repeatable severity filter
        required: false
        schema:
          type: array
          items:
            type: string
      - name: status
        in: query
        description: Repeatable status filter
        required: false
        schema:
          type: array
          items:
            type: string
      - name: source
        in: query
        description: Exact source filter
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Paginated alert list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertListResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
    post:
      tags:
      - alerts
      summary: POST /alerts — admin/maintainer only.
      operationId: create_alert
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBody'
        required: true
      responses:
        '201':
          description: Alert created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
  /api/v1/approvals:
    get:
      tags:
      - approvals
      summary: |-
        GET /approvals — any authenticated user; paginated with status[]/type[]
        (repeated) and requester_id filters, newest first.
      operationId: list_approvals
      parameters:
      - name: page
        in: query
        description: 1-based page number (default 1)
        required: false
        schema:
          type: integer
          format: int64
      - name: per_page
        in: query
        description: Page size, capped at 100 (default 20)
        required: false
        schema:
          type: integer
          format: int64
      - name: status
        in: query
        description: Repeatable status filter
        required: false
        schema:
          type: array
          items:
            type: string
      - name: type
        in: query
        description: Repeatable request_type filter
        required: false
        schema:
          type: array
          items:
            type: string
      - name: requester_id
        in: query
        description: Exact requester filter (0 = no filter)
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Paginated approval list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalListResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
    post:
      tags:
      - approvals
      summary: |-
        POST /approvals — any authenticated user; requester_id = current user,
        expires_at = now + expires_hours; 201 `{message, approval}`.
      operationId: create_approval
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBody'
        required: true
      responses:
        '201':
          description: Approval request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
  /api/v1/asm/scans:
    get:
      tags:
      - asm
      summary: GET /asm/scans — list ASM scans; query params forward first-value-wins.
      operationId: list_asm_scans
      responses:
        '200':
          description: Proxied scanner response
          content:
            application/json:
              schema: {}
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Cannot connect to scanner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: Scanner request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
    post:
      tags:
      - asm
      summary: POST /asm/scans — trigger a new ASM scan (proxied; body forwarded).
      description: |-
        This whole router is a thin authenticated proxy to the scanner service —
        every response body below is documented as a generic JSON object
        (`serde_json::Value`) rather than a fixed schema, since the wire shape is
        whatever the upstream scanner returns, not something this service
        defines. See `docs/v2-port/openapi-pattern.md`.
      operationId: create_asm_scan
      requestBody:
        content:
          application/json:
            schema: {}
        required: true
      responses:
        '200':
          description: Proxied scanner response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Cannot connect to scanner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: Scanner request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: GET /auth/me — the caller's own profile (mirrors `g.current_user`).
      operationId: me
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: |-
        POST /auth/refresh — unauthenticated (the refresh token in the body is
        the credential; no bearer header is required or checked).
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Rotated access/refresh token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid, revoked, or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/endpoint/agents:
    get:
      tags:
      - endpoint
      summary: |-
        GET /endpoint/agents — JWT, admin/maintainer. Filters: repeated `status` keys
        and a single `os_type`; ordered by last_heartbeat DESC; standard
        pagination envelope (per_page default 20, cap 100).
      operationId: list_agents
      parameters:
      - name: page
        in: query
        description: 1-based page number (default 1)
        required: false
        schema:
          type: integer
          format: int64
      - name: per_page
        in: query
        description: Page size, capped at 100 (default 20)
        required: false
        schema:
          type: integer
          format: int64
      - name: status
        in: query
        description: Repeatable status filter
        required: false
        schema:
          type: array
          items:
            type: string
      - name: os_type
        in: query
        description: Exact os_type filter
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Paginated agent list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
  /api/v1/endpoint/heartbeat:
    post:
      tags:
      - endpoint
      summary: |-
        POST /endpoint/heartbeat — HMAC agent auth. Updates status + last_heartbeat and
        merges the payload metadata over the stored metadata (new keys win),
        exactly like v1's `{**old, **new}`.
      operationId: heartbeat
      parameters:
      - name: X-Agent-ID
        in: header
        description: Agent identity — must match the HMAC key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeartbeatBody'
        required: true
      responses:
        '200':
          description: Heartbeat recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeartbeatResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing/invalid X-API-Key or X-Agent-ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Agent not registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - endpoint_hmac: []
...
  /api/v1/license/features:
    get:
      tags:
      - license
      summary: |-
        GET /license/features — tier, flag decisions, and feature entitlements
        for the authenticated caller's tenant; the frontend nav-gating contract.
      operationId: license_features
      responses:
        '200':
          description: License tier, flags, and features
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseFeaturesResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
  /api/v1/research/config:
    get:
      tags:
      - research
      summary: |-
        GET /research/config — research source status and timeouts. Config is read
        from the environment at request time (same pattern as alerts.rs `AI_ENABLED`
        and threat_intel.rs feed keys).
      operationId: get_config
      responses:
        '200':
          description: Research source status and timeouts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchConfigResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
  /api/v1/s3-scan/buckets:
    get:
      tags:
      - s3-scan
      summary: GET /buckets — paginated bucket configs, newest first, creds masked.
      operationId: list_buckets
      parameters:
      - name: page
        in: query
        description: 1-based page number (default 1)
        required: false
        schema:
          type: integer
          format: int64
      - name: per_page
        in: query
        description: Page size, capped at 500 (default 50)
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Paginated bucket configuration list (credentials masked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BucketListResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
    post:
      tags:
      - s3-scan
      summary: |-
        POST /buckets — admin/maintainer. 409 on duplicate endpoint+bucket pair.
        Defect #1: populates the NOT NULL created_by column v1 omitted.
      operationId: create_bucket
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BucketCreateBody'
        required: true
      responses:
        '201':
          description: Bucket configuration created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BucketCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Bucket configuration already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
  /api/v1/siem/config:
    get:
      tags:
      - siem
      summary: |-
        GET /siem/config — authenticated view of the current SIEM configuration
        (any role, matching v1's bare `@auth_required`).
      operationId: get_siem_config
      responses:
        '200':
          description: Current SIEM configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiemConfigResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
    put:
      tags:
      - siem
      summary: |-
        PUT /siem/config — admin-only retention update. Contract defect #5:
        v1 validates but never persists — replicated as a validate-only no-op.
        Invalid values return the bare v1 400 body, not the shared envelope.
      operationId: update_siem_config
      requestBody:
        content:
          application/json:
            schema: {}
        required: true
      responses:
        '200':
          description: 'Validated (not persisted — see contract defect #5)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiemConfigUpdateResponse'
        '400':
          description: retention_days must be an integer 1-400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Non-object JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
  /api/v1/users:
    get:
      tags:
      - users
      summary: GET /users — admin/maintainer only; paginated `{items,total,page,per_page,pages}`.
      operationId: list_users
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: per_page
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
    post:
      tags:
      - users
      summary: |-
        POST /users — admin only; 403 free-tier cap, 409 duplicate email,
        201 `{message,user}` on success.
      operationId: create_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequest'
        required: true
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions, or free-tier user cap reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
components:
  schemas:
...
    AlertCreateResponse:
      type: object
      description: Documentation-only mirror of `create_alert`'s `serde_json::json!` body.
      required:
      - message
      - alert
      properties:
        alert:
          $ref: '#/components/schemas/AlertCreateSummary'
        message:
          type: string
...
    AlertListResponse:
      type: object
      description: Documentation-only mirror of `list_alerts`'s `serde_json::json!` body.
      required:
      - items
      - total
      - page
      - per_page
      - pages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AlertItem'
        page:
          type: integer
          format: int64
        pages:
          type: integer
          format: int64
        per_page:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
...
    CreateBody:
      type: object
      description: |-
        ApprovalCreateRequest — fields optional here so missing ones map to the
        validation envelope instead of an axum extractor rejection.
      properties:
        expires_hours:
          type:
          - integer
          - 'null'
          format: int64
        metadata: {}
        request_type:
          type:
          - string
          - 'null'
        required_approvals:
          type:
          - integer
          - 'null'
          format: int64
        resource_id:
          type:
          - string
          - 'null'
        resource_type:
          type:
          - string
          - 'null'
...
    ErrorResponse:
      type: object
      description: |-
        Documentation-only mirror of `ApiError`'s bare `{"error": msg}` wire shape
        (`BadRequest`/`Unauthorized`/`Forbidden`/`NotFound`/`Conflict` in this
        service all use it, when their body happens to be bare) — `ApiError`
        itself builds its body with `serde_json::json!` rather than a typed
        struct, so this type exists solely to give `utoipa` something to
        reference in `#[utoipa::path]` `responses(...)` clauses. See
        `docs/v2-port/openapi-pattern.md`.
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable error message.
...
  securitySchemes:
    bearer_jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    endpoint_hmac:
      type: apiKey
      in: header
      name: X-API-Key
...